+++ ogma/src/common/mod.rs
@@ -1,0 +1,1 @@
+pub mod err;
