+++ ogma/src/eng/hir.rs
@@ -1,11 +1,8 @@
-use super::{
-    ast::{self, DefinitionImpl, Expression, Tag, Term},
-    impls::Implementation,
-    types::*,
-    var::*,
-    Definitions, Error, Result, Str,
-};
 use std::{convert::TryInto, fmt};
+use crate::prelude::*;
+use ast::{Tag, Expression, DefinitionImpl, Term};
+use crate::lang::var::{Locals, Environment};
+use Type as Ty;
 
 type StepR = Result<(Value, Environment)>;
 
@@ -18,7 +15,7 @@ pub fn handle_help(expr: &Expression, definitions: &Definitions) -> Result<()> {
             .iter()
             .any(|x| matches!(x, Term::Flag(f) if f.str() == "help"))
         {
-            Err(super::help_as_error(help))
+            Err(err::help_as_error(help))
         } else {
             Ok(())
         }
@@ -27,7 +24,7 @@ pub fn handle_help(expr: &Expression, definitions: &Definitions) -> Result<()> {
     }
 }
 
-pub(crate) fn construct_evaluator(
+pub fn construct_evaluator(
     in_ty: Type,
     expr: Expression,
     defs: &Definitions,
