+++ ogma/src/lang/mod.rs
@@ -1,0 +1,8 @@
+
+
+pub mod types;
+pub mod syntax;
+pub mod defs;
+pub mod impls;
+pub mod help;
+pub mod var;
