+++ ogma/src/lang/syntax/parse.rs
@@ -1,9 +1,5 @@
 //! Parsing source into a AST.
 
-use super::{
-    ast::{self, *},
-    Definitions, Error, ErrorTrace, HashSet,
-};
 use ::kserd::Number;
 use ::libs::divvy::Str;
 use nom::{
@@ -11,6 +7,7 @@ use nom::{
     error::*, multi::*, sequence::*, IResult, Offset,
 };
 use std::sync::Arc;
+use crate::prelude::{HashSet, err, ast::*, Definitions};
 
 struct Line {
     line: Arc<str>,
@@ -31,6 +28,30 @@ impl Line {
 }
 
 // ------ Entry and Error ------------------------------------------------------
+/// Successful parse result.
+pub enum ParseSuccess {
+    /// Parsed as a `def`inition.
+    Impl(DefinitionImpl),
+    /// Parsed as a type definition (`def-ty`).
+    Ty(DefinitionType),
+    /// Parsed as an expression.
+    Expr(Expression),
+}
+
+/// Parse the `input` as a valid `ogma` expression or definition.
+///
+/// Uses `Location::Shell`.
+pub fn parse(input: &str, defs: &Definitions) -> std::result::Result<ParseSuccess, ParseFail> {
+    let loc = Location::Shell;
+    if input.starts_with("def ") {
+        definition_impl(input, loc, defs).map(ParseSuccess::Impl)
+    } else if input.starts_with("def-ty ") {
+        definition_type(input, loc).map(ParseSuccess::Ty)
+    } else {
+        expression(input, loc, defs).map(ParseSuccess::Expr)
+    }
+}
+
 pub fn expression<S: Into<Arc<str>>>(
     expr: S,
     location: Location,
@@ -81,7 +102,7 @@ pub fn definition_type<S: Into<Arc<str>>>(
 }
 
 /// Failure to parse results in the parse [`Error`], and an expecting AST node.
-pub type ParseFail = (Error, Expecting);
+pub type ParseFail = (err::Error, Expecting);
 
 /// An indication of expected AST node type if parsing fails.
 ///
@@ -153,7 +174,7 @@ fn convert_parse_error<'a>(
     source: &'a str,
     loc: Location,
 ) -> ParseFail {
-    use crate::err::*;
+    use err::*;
     let ParsingError {
         input,
         cx,
@@ -174,7 +195,7 @@ fn convert_parse_error<'a>(
     let err = Error {
         cat: Category::Parsing,
         desc: "could not parse input line".into(),
-        traces: vec![ErrorTrace {
+        traces: vec![err::ErrorTrace {
             loc,
             source: source.into(),
             desc: Some(cx.to_string()),
@@ -280,11 +301,11 @@ fn expr<'f>(
 fn block<'f>(
     line: &'f Line,
     defs: &'f Definitions,
-) -> impl for<'a> Fn(&'a str) -> IResult<&'a str, ast::PrefixBlock, ParsingError<'a>> + 'f {
+) -> impl for<'a> Fn(&'a str) -> IResult<&'a str, PrefixBlock, ParsingError<'a>> + 'f {
     move |i| {
         let (i, op) = exp(op(line), Expecting::Impl)(i)?;
         let (i, terms) = many0(ws(term(line, defs)))(i)?;
-        Ok((i, ast::PrefixBlock { op, terms }))
+        Ok((i, PrefixBlock { op, terms }))
     }
 }
 
@@ -773,9 +794,9 @@ mod tests {
     ) -> Result<T, ParseFail> {
         Err((
             Error {
-                cat: crate::err::Category::Parsing,
+                cat: err::Category::Parsing,
                 desc: "could not parse input line".into(),
-                traces: vec![ErrorTrace {
+                traces: vec![err::ErrorTrace {
                     loc: Location::Shell,
                     source: source.into(),
                     desc: Some(cx.into()),
@@ -1670,7 +1691,7 @@ mod tests {
     #[test]
     fn incomplete_expecting_tests() {
         let exp = |s| {
-            crate::parse(s, &Definitions::default())
+            parse(s, &Definitions::default())
                 .map(|_| ())
                 .unwrap_err()
                 .1
