+++ ogma/src/rt/bat.rs
@@ -2,9 +2,6 @@
 //!
 //! This module provides some structure around batch processing, supplying reporting structures
 //! and evaluation semantics.
-use crate::{
-    process_definition, process_expression, recognise_definition, Error, Location, Mutex, Result,
-};
 use ::libs::{
     divvy::*,
     rayon::prelude::*,
@@ -18,6 +15,9 @@ use std::{
     sync::Arc,
     time::{Duration, Instant},
 };
+use crate::prelude::*;
+use crate::Mutex;
+use ast::Location;
 
 /// A batch to process.
 pub struct Batch {
@@ -101,7 +101,7 @@ pub enum ItemType {
 
 fn work_out_type(code: &str) -> ItemType {
     let c = code.trim_start();
-    if recognise_definition(c) {
+    if lang::defs::recognise_definition(c) {
         if c.starts_with("def-ty") {
             ItemType::Type
         } else {
@@ -261,7 +261,7 @@ pub fn process(
     root: &Path,
     wd: &Path,
     progress: &ProgressTx,
-    mut definitions: crate::Definitions,
+    mut definitions: Definitions,
 ) -> Vec<(Outcome, Duration)> {
     let ff = batch.fail_fast;
     let parallelise = batch.parallelise;
@@ -316,7 +316,7 @@ pub fn process(
 
         let instant = Instant::now();
         let loc = Location::File(def.file.clone(), def.line as usize);
-        let r = process_definition(&def.code, loc, def.comment.clone(), &mut definitions).into();
+        let r = lang::defs::process_definition(&def.code, loc, def.comment.clone(), &mut definitions).into();
         report_progress(prog, &reporter, idx, &r);
         sw_if_fail(&r);
         results[idx] = (r, instant.elapsed());
@@ -338,7 +338,7 @@ pub fn process(
         }
 
         let loc = Location::File(expr.file.clone(), expr.line as usize);
-        let r = process_expression((), expr.code.as_str(), loc, &definitions, root, wd).into();
+        let r = rt::process_expression((), expr.code.as_str(), loc, &definitions, root, wd).into();
         report_progress(prog, &reporter, idx, &r);
         sw_if_fail(&r);
         (idx, r, instant.elapsed())
